split out get function
